wrong cast for 64 bit targets
git-svn-id: svn://svn.dd-wrt.com/DD-WRT@27474 52c4871e-980c-0410-b1e0-e73912ce01f8

Former-commit-id: bf743fafdf31da760eeebe61c48ea6e6f27e13e4

+++ src/router/upnp/src/gena.c
@@ -261,7 +261,7 @@ notify_prop_change(UPNP_CONTEXT *context, UPNP_SUBSCRIBER *subscriber)
 
 	/* set non-blocking IO to have connect() return without pending */
 	on = 0;
-	if ((ioctl(s, FIONBIO, (int)&on)) == -1) {
+	if ((ioctl(s, FIONBIO, &on)) == -1) {
 		upnp_syslog(LOG_ERR, "cannot set socket as blocking IO");
 		goto error_out;
 	}
